# backend/docs/swagger.yaml
definitions:
...
  request.NewInvitation:
    properties:
      bypassApproval:
        type: boolean
      expiresAt:
        type: string
      joinLimit:
        type: integer
    type: object
  request.NewServer:
    properties:
      name:
        example: My very good server
        maxLength: 256
        type: string
    required:
    - name
    type: object
...
  request.UpdateServer:
    properties:
      announcementChannel:
        type: string
      bannerUrl:
        type: string
      defaultPermission:
        type: string
      description:
        type: string
      iconUrl:
        type: string
      name:
        type: string
      needApproval:
        type: boolean
    type: object
  response.Channel:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      parentCategory:
        type: string
      serverId:
        type: string
      updatedAt:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.GetServerResponse:
    properties:
      announcementChannel:
        type: string
      bannerUrl:
        type: string
      channels:
        items:
          $ref: '#/definitions/response.Channel'
        type: array
      createdAt:
        type: string
      defaultRole:
        type: string
      description:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/response.Role'
        type: array
      selfMembership:
        $ref: '#/definitions/response.Membership'
      updatedAt:
        type: string
    type: object
  response.GetServersResponse:
    properties:
      result:
        items:
          $ref: '#/definitions/response.ServerPreview'
        type: array
    type: object
...
  response.Invitation:
    properties:
      bypassApproval:
        type: boolean
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      joinCount:
        type: integer
      joinLimit:
        type: integer
      serverId:
        type: string
    type: object
...
  response.Membership:
    properties:
      assignedRoles:
        items:
          type: string
        type: array
      createdAt:
        type: string
      nickname:
        type: string
      serverId:
        type: string
      userId:
        type: string
    type: object
...
  response.NewServerResponse:
    properties:
      id:
        type: string
    type: object
  response.Role:
    properties:
      allowMention:
        type: boolean
      color:
        type: integer
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      priority:
        type: integer
      serverId:
        type: string
    type: object
  response.ServerPreview:
    properties:
      bannerUrl:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  response.UpdateServerResponse:
    properties:
      announcementChannel:
        type: string
      bannerUrl:
        type: string
      createdAt:
        type: string
      defaultRole:
        type: string
      description:
        type: string
      iconUrl:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: This is the api for Noncord
  title: Noncord API
  version: "1.0"
paths:
...
  /api/v1/server:
    get:
      description: Get all servers the user is in
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetServersResponse'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get servers by user
      tags:
      - Server
    post:
      consumes:
      - application/json
      description: Create a server
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data for creating server
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.NewServer'
      produces:
      - application/json
      responses:
        "200":
          description: Server id
          schema:
            $ref: '#/definitions/response.NewServerResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a server
      tags:
      - Server
  /api/v1/server/{server_id}:
    delete:
      description: Delete server by id
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server Id
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete server
      tags:
      - Server
    get:
      description: Get a server by id, can only get server the user is in
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: server id to fetch
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetServerResponse'
        "400":
          description: server_id not a valid id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a server
      tags:
      - Server
    patch:
      consumes:
      - application/json
      description: Update server
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update server
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.UpdateServer'
      - description: Server Id
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated server
          schema:
            $ref: '#/definitions/response.UpdateServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update server
      tags:
      - Server
    put:
      consumes:
      - application/json
      description: Update server
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update server
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.UpdateServer'
      - description: Server Id
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated server
          schema:
            $ref: '#/definitions/response.UpdateServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update server
      tags:
      - Server
  /api/v1/server/{server_id}/invitations:
    get:
      description: Get all server's invitations
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server Id
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Invitation'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get server's invitations
      tags:
      - Server
    post:
      description: Get an invitation on a server
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server Id
        in: path
        name: server_id
        required: true
        type: integer
      - description: Data for creating invitation
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.NewInvitation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Invitation'
        "401":
          description: Unknown session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Server not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create invitation
      tags:
      - Server
...
swagger: "2.0"